+++ flutter_navigation_generator/test/code_builder/on_generate_route_builder_test.dart
@@ -24,27 +24,15 @@ void main() {
       expect(content.requiredParameters.first.type?.symbol, 'RouteSettings');
       expect(content.requiredParameters.length, 1);
       expect(content.body is Block, true);
-      expect(
-        (content.body as Block).statements.first.toString(),
-        'return null;',
-      );
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
     test('1 test route no return', () {
-      final content =
-          OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-              ),
-            },
-          ).generate();
+      final content = OnGenerateRouteBuilder(routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'))}).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -61,25 +49,17 @@ void main() {
       final content =
           OnGenerateRouteBuilder(
             routes: {
-              RouteConfig(
-                routeName: 'test_without_return',
-                routeWidget: const ImportableType(
-                  className: 'TestPageWithoutReturn',
-                ),
-              ),
+              RouteConfig(routeName: 'test_without_return', routeWidget: const ImportableType(className: 'TestPageWithoutReturn')),
               RouteConfig(
                 routeName: 'test_with_return',
-                routeWidget: const ImportableType(
-                  className: 'TestPageWithReturn',
-                ),
+                routeWidget: const ImportableType(className: 'TestPageWithReturn'),
                 returnType: const ImportableType(className: 'ReturnType'),
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -97,25 +77,17 @@ void main() {
           OnGenerateRouteBuilder(
             unknownRoute: const ImportableType(className: 'Route404'),
             routes: {
-              RouteConfig(
-                routeName: 'test_without_return',
-                routeWidget: const ImportableType(
-                  className: 'TestPageWithoutReturn',
-                ),
-              ),
+              RouteConfig(routeName: 'test_without_return', routeWidget: const ImportableType(className: 'TestPageWithoutReturn')),
               RouteConfig(
                 routeName: 'test_with_return',
-                routeWidget: const ImportableType(
-                  className: 'TestPageWithReturn',
-                ),
+                routeWidget: const ImportableType(className: 'TestPageWithReturn'),
                 returnType: const ImportableType(className: 'ReturnType'),
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -129,19 +101,10 @@ void main() {
     });
 
     test('test routename starting with a number', () {
-      final content =
-          OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: '404',
-                routeWidget: const ImportableType(className: 'Route404'),
-              ),
-            },
-          ).generate();
+      final content = OnGenerateRouteBuilder(routes: {RouteConfig(routeName: '404', routeWidget: const ImportableType(className: 'Route404'))}).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.r404: return MaterialPageRoute<void>(builder: (_) => Route404(), settings: settings, fullscreenDialog: false,);}",
@@ -156,19 +119,10 @@ void main() {
 
     test('1 test route with constructorName', () {
       final content =
-          OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                constructorName: 'constructorName',
-              ),
-            },
-          ).generate();
+          OnGenerateRouteBuilder(routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), constructorName: 'constructorName')}).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage.constructorName(), settings: settings, fullscreenDialog: false,);}",
@@ -183,19 +137,10 @@ void main() {
 
     test('1 test route with isFullscreenDialog', () {
       final content =
-          OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                isFullscreenDialog: true,
-              ),
-            },
-          ).generate();
+          OnGenerateRouteBuilder(routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), isFullscreenDialog: true)}).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: true,);}",
@@ -211,18 +156,11 @@ void main() {
     test('1 test route with CustomPageType', () {
       final content =
           OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                pageType: const ImportableType(className: 'CustomPageType'),
-              ),
-            },
+            routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), pageType: const ImportableType(className: 'CustomPageType'))},
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -238,18 +176,12 @@ void main() {
     test('1 test route with CustomPageType from navigator', () {
       final content =
           OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-              ),
-            },
+            routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'))},
             pageType: const ImportableType(className: 'CustomPageType'),
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -271,34 +203,20 @@ void main() {
                 routeWidget: const ImportableType(className: 'TestPage'),
                 parameters: [
                   RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'bool',
-                      name: 'testBool',
-                      isNullable: true,
-                    ),
+                    type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
                     addToJson: true,
                     defaultValue: null,
                     ignore: false,
                     queryName: 'testBool',
                   ),
-                  RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'int',
-                      name: 'testInt',
-                    ),
-                    addToJson: true,
-                    defaultValue: null,
-                    ignore: false,
-                    queryName: 'testInt',
-                  ),
+                  RouteFieldConfig(type: const ImportableType(className: 'int', name: 'testInt'), addToJson: true, defaultValue: null, ignore: false, queryName: 'testInt'),
                 ],
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: queryParameters['testBool'] != null ? queryParameters['testBool']! == 'true' : arguments['testBool'] as bool?,testInt: queryParameters['testInt'] != null ? int.parse(queryParameters['testInt']!) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -320,34 +238,20 @@ void main() {
                 routeWidget: const ImportableType(className: 'TestPage'),
                 parameters: [
                   RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'bool',
-                      name: 'testBool',
-                      isNullable: true,
-                    ),
+                    type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
                     addToJson: true,
                     defaultValue: null,
                     ignore: false,
                     queryName: 'testBool',
                   ),
-                  RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'int',
-                      name: 'testInt',
-                    ),
-                    addToJson: true,
-                    defaultValue: null,
-                    ignore: false,
-                    queryName: 'testInt',
-                  ),
+                  RouteFieldConfig(type: const ImportableType(className: 'int', name: 'testInt'), addToJson: true, defaultValue: null, ignore: false, queryName: 'testInt'),
                 ],
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {}",
@@ -372,34 +276,20 @@ void main() {
                 routeWidget: const ImportableType(className: 'TestPage'),
                 parameters: [
                   RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'bool',
-                      name: 'testBool',
-                      isNullable: true,
-                    ),
+                    type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
                     addToJson: true,
                     defaultValue: null,
                     ignore: false,
                     queryName: 'testBool',
                   ),
-                  RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'int',
-                      name: 'testInt',
-                    ),
-                    addToJson: true,
-                    defaultValue: null,
-                    ignore: false,
-                    queryName: 'testInt',
-                  ),
+                  RouteFieldConfig(type: const ImportableType(className: 'int', name: 'testInt'), addToJson: true, defaultValue: null, ignore: false, queryName: 'testInt'),
                 ],
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.testPage: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: queryParameters['testBool'] != null ? queryParameters['testBool']! == 'true' : arguments['testBool'] as bool?,testInt: queryParameters['testInt'] != null ? int.parse(queryParameters['testInt']!) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -421,39 +311,25 @@ void main() {
                 routeWidget: const ImportableType(className: 'TestPage'),
                 parameters: [
                   RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'bool',
-                      name: 'testBool',
-                      isNullable: true,
-                    ),
+                    type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
                     addToJson: true,
                     defaultValue: null,
                     ignore: false,
                     queryName: 'testBool',
                   ),
-                  RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'int',
-                      name: 'testInt',
-                    ),
-                    addToJson: true,
-                    defaultValue: null,
-                    ignore: false,
-                    queryName: 'testInt',
-                  ),
+                  RouteFieldConfig(type: const ImportableType(className: 'int', name: 'testInt'), addToJson: true, defaultValue: null, ignore: false, queryName: 'testInt'),
                 ],
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {}",
         "final pathSegments = settingsUri.pathSegments;",
-        "if (pathSegments.length == 3) {if (pathSegments[0] == 'test' && pathSegments[2] == '') {queryParameters['testBool'] = pathSegments[1];return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: queryParameters['testBool'] != null ? queryParameters['testBool']! == 'true' : arguments['testBool'] as bool?,testInt: queryParameters['testInt'] != null ? int.parse(queryParameters['testInt']!) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}}",
+        "if (pathSegments.length == 2) {if (pathSegments[0] == 'test') {queryParameters['testBool'] = pathSegments[1];return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: queryParameters['testBool'] != null ? queryParameters['testBool']! == 'true' : arguments['testBool'] as bool?,testInt: queryParameters['testInt'] != null ? int.parse(queryParameters['testInt']!) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}}",
         "return null;",
       ];
       if (printActuals) print(jsonEncode(bodyStatements));
@@ -472,34 +348,20 @@ void main() {
                 routeWidget: const ImportableType(className: 'TestPage'),
                 parameters: [
                   RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'bool',
-                      name: 'testBool',
-                      isNullable: true,
-                    ),
+                    type: const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
                     addToJson: true,
                     defaultValue: 'true',
                     ignore: false,
                     queryName: 'testBool',
                   ),
-                  RouteFieldConfig(
-                    type: const ImportableType(
-                      className: 'int',
-                      name: 'testInt',
-                    ),
-                    addToJson: true,
-                    defaultValue: '1',
-                    ignore: false,
-                    queryName: 'testInt',
-                  ),
+                  RouteFieldConfig(type: const ImportableType(className: 'int', name: 'testInt'), addToJson: true, defaultValue: '1', ignore: false, queryName: 'testInt'),
                 ],
               ),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: queryParameters['testBool'] != null ? queryParameters['testBool']! == 'true' : arguments['testBool'] as bool? ?? true,testInt: queryParameters['testInt'] != null ? int.parse(queryParameters['testInt']!) : arguments['testInt'] as int? ?? 1,), settings: settings, fullscreenDialog: false,);}",
@@ -514,61 +376,32 @@ void main() {
 
     test('1 test route with generatePageRoute false', () {
       final content =
-          OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                generatePageRoute: false,
-              ),
-            },
-          ).generate();
+          OnGenerateRouteBuilder(routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), generatePageRoute: false)}).generate();
 
       expect(content.body is Block, true);
-      expect(
-        (content.body as Block).statements.first.toString(),
-        'return null;',
-      );
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
     test('1 test route with dialog', () {
       final content =
           OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                navigationType: NavigationType.dialog,
-              ),
-            },
+            routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), navigationType: NavigationType.dialog)},
           ).generate();
 
       expect(content.body is Block, true);
-      expect(
-        (content.body as Block).statements.first.toString(),
-        'return null;',
-      );
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
     test('1 test route with bottomsheet', () {
       final content =
           OnGenerateRouteBuilder(
-            routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                navigationType: NavigationType.bottomSheet,
-              ),
-            },
+            routes: {RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), navigationType: NavigationType.bottomSheet)},
           ).generate();
 
       expect(content.body is Block, true);
-      expect(
-        (content.body as Block).statements.first.toString(),
-        'return null;',
-      );
+      expect((content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -576,21 +409,13 @@ void main() {
       final content =
           OnGenerateRouteBuilder(
             routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-                guards: const [ImportableType(className: 'TestGuard')],
-              ),
-              RouteConfig(
-                routeName: 'test2',
-                routeWidget: const ImportableType(className: 'TestPage2'),
-              ),
+              RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage'), guards: const [ImportableType(className: 'TestGuard')]),
+              RouteConfig(routeName: 'test2', routeWidget: const ImportableType(className: 'TestPage2')),
             },
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",
@@ -607,22 +432,14 @@ void main() {
       final content =
           OnGenerateRouteBuilder(
             routes: {
-              RouteConfig(
-                routeName: 'test',
-                routeWidget: const ImportableType(className: 'TestPage'),
-              ),
-              RouteConfig(
-                routeName: 'test2',
-                routeWidget: const ImportableType(className: 'TestPage2'),
-                guards: [],
-              ),
+              RouteConfig(routeName: 'test', routeWidget: const ImportableType(className: 'TestPage')),
+              RouteConfig(routeName: 'test2', routeWidget: const ImportableType(className: 'TestPage2'), guards: []),
             },
             defaultGuards: [const ImportableType(className: 'TestGuard')],
           ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements =
-          (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    final queryParameters = Map.from(settingsUri.queryParameters);",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",